requirements.txt
flask
google-cloud-firestore
firebase-admin
google-cloud-storage
python-dotenv
langchain
openai
chromadb
tiktoken
langchain-community